Ajustes diversos na geração dos vizinhos na busca local. Resta investigar o funcionamento dos demais loops externos do GRASP.

+++ mestrado/src/problem/RCCProblem.cpp
@@ -22,8 +22,8 @@ RCCProblem::~RCCProblem() {
 /**
  * Returns he Relaxed Imbalance of a partition P (RI(P)).
  */
-int RCCProblem::objectiveFunction(SignedGraph* g, Clustering *c) {
-	return 0;
+float RCCProblem::objectiveFunction(SignedGraph* g, Clustering *c) {
+	return 0.0;
 }
 
 } /* namespace problem */
